refactor: viewholder, 패키지 분리
