+++ app/src/main/java/woowacourse/paint/adapter/ColorPaletteViewHolder.kt
@@ -1,0 +1,19 @@
+package woowacourse.paint.adapter
+
+import androidx.recyclerview.widget.RecyclerView
+import woowacourse.paint.brush.ColorPalette
+import woowacourse.paint.databinding.ItemColorPaletteBinding
+
+class ColorPaletteViewHolder(
+    private val binding: ItemColorPaletteBinding,
+    private val onClickColorPalette: (ColorPalette) -> Unit,
+) : RecyclerView.ViewHolder(binding.root) {
+    fun bind(colorPalette: ColorPalette) {
+        binding.paletteButton.apply {
+            setBackgroundColor(colorPalette.color)
+            setOnClickListener {
+                onClickColorPalette(colorPalette)
+            }
+        }
+    }
+}
